refactoring
- EnableNotificationsBtn is a new components
- only shows up if there is service worker and browser supports EnableNotificationsBtn
- invisible on click
